fix(laravel): conditional return type for phpstan
This adds a conditional return type annotation [0] for phpstan.
With this annotation, projects with phpstan enabled will not see errors
anymore when using the hybridly() helper function.

Additionally we only use the phpstan specific annotation, to avoid
unnecessary issues with tools like intelliphense, which do not
yet support conditional types.

[0] https://phpstan.org/writing-php-code/phpdoc-types#conditional-return-types