feat(routing): Add pattern matching for current
Also adds the functionality to get the current route name, when calling
the function without parameters.

The implementation heavily inspired by ziggy and is mainly built around
regex matching and will transform the route definitions to regexes which
are then used to check against the current URL.

Additionally some test cases are provided, but requires more realworld
testing.

Also note that the mode parameter was removed from the current function
as it should not be needed anymore.