ci: Add provenance to npm builds
This commit adds provenance for all published packages. See the NPM documentation [0].

Provenance will allow people to verify that the hybridly packages were actually built on GH Actions and with the content of the corresponding commit. This will help with supply chain security.

For this to work, the `id-token` permission was added only where necessary.

[0]: https://docs.npmjs.com/generating-provenance-statements